+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Ver 0.0.3
+2015-06-02
+* consolidated statistics in class and added weekly email
+
 2015-05-28
 * added statistics on returning cars (i.e. anonymous users)